# docs/swagger.yaml
definitions:
...
  dto.Resp:
    properties:
      cate_id:
        type: integer
      code:
        type: integer
      cur_page:
        type: integer
      data:
        type: object
      msg:
        type: string
      total:
        type: integer
    type: object
...
  go_micro_service_novel.Category:
    properties:
      cate_id:
        type: integer
      channel:
        type: integer
      name:
        type: string
      sort:
        type: integer
    type: object
  go_micro_service_novel.Chapter:
    properties:
      chapter_id:
        type: integer
      content:
        type: string
      is_vip:
        type: integer
      novel_id:
        type: integer
      num:
        type: integer
      title:
        type: string
      words:
        type: integer
    type: object
  go_micro_service_novel.Note:
    properties:
      new_chapter:
        type: string
      new_num:
        type: integer
      novel_id:
        type: integer
      novel_name:
        type: string
      prev_chapter:
        type: string
      prev_num:
        type: integer
    type: object
  go_micro_service_novel.Novel:
    properties:
      author:
        type: string
      cate_id:
        type: integer
      chapter_current:
        type: integer
      chapter_total:
        type: integer
      img:
        type: string
      intro:
        type: string
      likes:
        type: integer
      name:
        type: string
      new_chapter:
        type: string
      novel_id:
        type: integer
      un_likes:
        type: integer
      updated_at:
        type: string
      words:
        type: integer
    type: object
...
info:
  contact:
    email: scg130@163.com
    name: scg130
  description: micro
  title: micro
  version: "1.0"
paths:
...
  /novel/cates:
    get:
      description: 分类列表
      parameters:
      - in: query
        name: cate_id
        type: integer
      - in: query
        name: chapter_id
        type: integer
      - in: query
        name: is_end
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: novel_id
        type: integer
      - in: query
        name: num
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: size
        type: integer
      - in: query
        name: type
        type: string
      - in: query
        name: words
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/go_micro_service_novel.Category'
                  type: array
              type: object
      summary: 分类列表
      tags:
      - novel
  /novel/chapter:
    get:
      description: 获取章节
      parameters:
      - description: query参数
        in: query
        name: chapter_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  $ref: '#/definitions/go_micro_service_novel.Chapter'
              type: object
      summary: 获取章节
      tags:
      - novel
  /novel/chapters:
    get:
      description: 章节列表
      parameters:
      - in: query
        name: cate_id
        type: integer
      - in: query
        name: chapter_id
        type: integer
      - in: query
        name: is_end
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: novel_id
        type: integer
      - in: query
        name: num
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: size
        type: integer
      - in: query
        name: type
        type: string
      - in: query
        name: words
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/go_micro_service_novel.Chapter'
                  type: array
              type: object
      summary: 章节列表
      tags:
      - novel
  /novel/join-book:
    get:
      description: 我的书架
      parameters:
      - in: query
        name: cate_id
        type: integer
      - in: query
        name: chapter_id
        type: integer
      - in: query
        name: is_end
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: novel_id
        type: integer
      - in: query
        name: num
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: size
        type: integer
      - in: query
        name: type
        type: string
      - in: query
        name: words
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/go_micro_service_novel.Note'
                  type: array
              type: object
      summary: 加入书架
      tags:
      - novel
  /novel/list:
    get:
      description: 小说列表
      parameters:
      - in: query
        name: cate_id
        type: integer
      - in: query
        name: chapter_id
        type: integer
      - in: query
        name: is_end
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: novel_id
        type: integer
      - in: query
        name: num
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: size
        type: integer
      - in: query
        name: type
        type: string
      - in: query
        name: words
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/go_micro_service_novel.Novel'
                  type: array
              type: object
      summary: 小说列表
      tags:
      - novel
  /novel/notes:
    get:
      description: 我的书架
      parameters:
      - in: query
        name: cate_id
        type: integer
      - in: query
        name: chapter_id
        type: integer
      - in: query
        name: is_end
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: novel_id
        type: integer
      - in: query
        name: num
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: size
        type: integer
      - in: query
        name: type
        type: string
      - in: query
        name: words
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/go_micro_service_novel.Note'
                  type: array
              type: object
      summary: 我的书架
      tags:
      - novel
  /novel/novel:
    get:
      description: 获取小说详情
      parameters:
      - description: query参数
        in: query
        name: novel_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  $ref: '#/definitions/go_micro_service_novel.Novel'
              type: object
      summary: 获取小说详情
      tags:
      - novel
  /novel/search:
    get:
      description: 搜索
      parameters:
      - in: query
        name: cate_id
        type: integer
      - in: query
        name: chapter_id
        type: integer
      - in: query
        name: is_end
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: novel_id
        type: integer
      - in: query
        name: num
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: size
        type: integer
      - in: query
        name: type
        type: string
      - in: query
        name: words
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/go_micro_service_novel.Novel'
                  type: array
              type: object
      summary: 搜索
      tags:
      - novel
...
schemes:
- http
swagger: "2.0"